new Function: Recover deleted Items

+++ books/view.js
@@ -2,22 +2,52 @@ var template = [{id: 0, title: '', author: '', deleted: 0}];
 
 let table = '';
 
+let deletedButton = '<a href="/books/deleted"><button class="deletedButton">gelöschte Bücher</button></a>';
+let activeButton = '<a href="/books/"><button class="activeButton">aktive Bücher</button></a>'
+
+let deletedHeadline = 'Gelöschte Bücher';
+let activeHeadline = 'Aktive Bücher';
+
 module.exports = { 
     render(books, deleted) {
+
+        let button = '';
+        let headline = '';
+        if(deleted === 1) {
+            button = activeButton;
+            headline = deletedHeadline;
+        } else {
+            button = deletedButton;
+            headline = activeHeadline;
+        }
+
         template = books;
         table = '';
         for(let i = 0; i < template.length; i++) {
-            if(books[i].deleted === deleted){
-                let row =
-                    '<tr>' +
-                        '<td>' + template[i].id + '</td>' +
-                        '<td>' + template[i].title + '</td>' +
-                        '<td>' + template[i].author + '</td>' +
-                        '<td><a href="/books/delete/' + template[i].id + '"><button class="deleteButton">löschen</button></a></td>' +
-                        '<td><a href="/books/form/' + template[i].id + '"><button class="editButton">bearbeiten</button></a></td>' + 
-                    '</tr>';
-                table += row;
-            }  
+            if(books[i].deleted === deleted) {
+                if(books[i].deleted === 0){
+                    let row =
+                        '<tr>' +
+                            '<td>' + template[i].id + '</td>' +
+                            '<td>' + template[i].title + '</td>' +
+                            '<td>' + template[i].author + '</td>' +
+                            '<td><a href="/books/delete/' + template[i].id + '"><button class="deleteButton">löschen</button></a></td>' +
+                            '<td><a href="/books/form/' + template[i].id + '"><button class="editButton">bearbeiten</button></a></td>' + 
+                        '</tr>';
+                    table += row;
+                }  else {
+                    let row =
+                        '<tr>' +
+                            '<td>' + template[i].id + '</td>' +
+                            '<td>' + template[i].title + '</td>' +
+                            '<td>' + template[i].author + '</td>' +
+                            '<td><a href="/books/recover/' + template[i].id + '"><button class="recoverButton">wiederherstellen</button></a></td>' +
+                            '<td><a href="/books/form/' + template[i].id + '"><button class="editButton">bearbeiten</button></a></td>' + 
+                        '</tr>';
+                    table += row;
+                }
+            }
+            
         }
 
         return `
@@ -29,6 +59,7 @@ module.exports = {
                 <link rel="stylesheet" href="/stylesheets/books.css" />
             </head>
             <body>
+                <h1 class="headLine">${headline}</h1>
                 <table>
                     <thead>
                         <tr><th>ID in database</th><th>Title</th><th>Author</th><th></th><th></th></tr>
@@ -40,7 +71,8 @@ module.exports = {
                 <div class="siteBox">
                     <div class="optionsBox">
                         <h1>Bücher</h1>
-                        <a href="/books/form"><button class="newButton">neu</button></a>
+                        <a href="/books/form"><button class="newButton">neues Buch</button></a>
+                        ${button}
                     </div>
                     <div class="infobox">
                         <h2>&copy Tim Gabrikowski</h2>
